settings use empty string if environment variables are missing
we use the environment variables by default, but if they are not set
we can just use empty strings. Of course in production empty strings
are not going to work, but we give the opportunity for the settings
to be set to the settings module in runtime, specially helpful
for tests.