internalusers.create_password accepts properties (API request body)
So the caller can specify roles or other attributes, and it's possible
to specify the password hash (instead of the password itself).

Since the password itslef is one of the acceptable properties, it's
possible to pass the password as one of the properties, but to avoid
ambiguity, if the password is also defined explicitly (not required)
and it does not match the specified password in properties, a ValueError
is raised.
So the password can be specified:

 - Just an explicit argument for password
 - Only as a property, with no explicit password argument
 - Both as explicit password argument and a property, but only if they
 match