Changes to settings are visible to roles/internalusers modules
We were using "from settings import" style importing for settings (API
URL and AUTH) in other modules (internalusers/roles).
That caused the settings to be copied to the modules (the settinsg
are string values), changes later to the settings were not affecting
the module behaviors.

With this commit those modules use FQN when accessing the settings,
so the changes to the settings is going to be effective.

Also replaced mock.ANY to ANY in tests and remove "import mock as mock".